migrate part of proposal to simapp
